worker-service/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.2
sqlalchemy==2.0.36
boto3==1.35.24
prometheus-client